+++ yggdrasil_decision_forests/port/python/CHANGELOG.md
@@ -1,12 +1,16 @@
 # Changelog
 
-## Head
+## HEAD
 
 ### Feature
 
 -   Warn if training on a VerticalDataset and fail if attempting to modify the
     columns in a VerticalDataset during training.
 -   User can override the model's task, label or group during evaluation.
+
+### Fix
+
+-   Remove the unsupported / invalid hyperparameters from isolation forests.
 
 ## 0.7.0 - 2024-08-21